Add --cache-downloads for Firmwares

+++ Mist/Commands/Download/DownloadFirmwareCommand.swift
@@ -68,6 +68,7 @@ struct DownloadFirmwareCommand: ParsableCommand {
 
         !options.quiet ? PrettyPrint.print("Include betas in search results will be '\(options.includeBetas)'...") : Mist.noop()
         !options.quiet ? PrettyPrint.print("Only include compatible firmwares will be '\(options.compatible)'...") : Mist.noop()
+        !options.quiet ? PrettyPrint.print("Cache downloads will be '\(options.cacheDownloads)'...") : Mist.noop()
         !options.quiet ? PrettyPrint.print("Output directory will be '\(options.outputDirectory)'...") : Mist.noop()
         !options.quiet ? PrettyPrint.print("Temporary directory will be '\(options.temporaryDirectory)'...") : Mist.noop()
         !options.quiet ? PrettyPrint.print("Force flag\(options.force ? " " : " has not been ")set, existing files will\(options.force ? " " : " not ")be overwritten...") : Mist.noop()
@@ -138,21 +139,31 @@ struct DownloadFirmwareCommand: ParsableCommand {
 
         let outputURL: URL = URL(fileURLWithPath: outputDirectory(for: firmware, options: options))
         let temporaryURL: URL = URL(fileURLWithPath: temporaryDirectory(for: firmware, options: options))
+        var processing: Bool = false
 
         !options.quiet ? PrettyPrint.printHeader("SETUP") : Mist.noop()
 
         if !FileManager.default.fileExists(atPath: outputURL.path) {
             !options.quiet ? PrettyPrint.print("Creating output directory '\(outputURL.path)'...") : Mist.noop()
             try FileManager.default.createDirectory(atPath: outputURL.path, withIntermediateDirectories: true, attributes: nil)
+            processing = true
         }
 
-        if FileManager.default.fileExists(atPath: temporaryURL.path) {
+        if FileManager.default.fileExists(atPath: temporaryURL.path) && !options.cacheDownloads {
             !options.quiet ? PrettyPrint.print("Deleting old temporary directory '\(temporaryURL.path)'...") : Mist.noop()
             try FileManager.default.removeItem(at: temporaryURL)
+            processing = true
         }
 
-        !options.quiet ? PrettyPrint.print("Creating new temporary directory '\(temporaryURL.path)'...") : Mist.noop()
-        try FileManager.default.createDirectory(at: temporaryURL, withIntermediateDirectories: true, attributes: nil)
+        if !FileManager.default.fileExists(atPath: temporaryURL.path) {
+            !options.quiet ? PrettyPrint.print("Creating new temporary directory '\(temporaryURL.path)'...") : Mist.noop()
+            try FileManager.default.createDirectory(at: temporaryURL, withIntermediateDirectories: true, attributes: nil)
+            processing = true
+        }
+
+        if !processing {
+            !options.quiet ? PrettyPrint.print("Nothing to do!") : Mist.noop()
+        }
     }
 
     /// Verifies free space for macOS Firmware downloads.
@@ -191,13 +202,14 @@ struct DownloadFirmwareCommand: ParsableCommand {
     ///
     /// - Throws: An `Error` if any of the directory operations fail.
     private static func teardown(_ firmware: Firmware, options: DownloadFirmwareOptions) throws {
-
         let temporaryURL: URL = URL(fileURLWithPath: temporaryDirectory(for: firmware, options: options))
+        !options.quiet ? PrettyPrint.printHeader("TEARDOWN") : Mist.noop()
 
-        if FileManager.default.fileExists(atPath: temporaryURL.path) {
-            !options.quiet ? PrettyPrint.printHeader("TEARDOWN") : Mist.noop()
+        if FileManager.default.fileExists(atPath: temporaryURL.path) && !options.cacheDownloads {
             !options.quiet ? PrettyPrint.print("Deleting temporary directory '\(temporaryURL.path)'...", prefix: .ending) : Mist.noop()
             try FileManager.default.removeItem(at: temporaryURL)
+        } else {
+            !options.quiet ? PrettyPrint.print("Nothing to do!", prefix: options.exportPath != nil ? .default : .ending) : Mist.noop()
         }
     }
